portfolio/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-taggit==6.1.0
django-unfold==0.60.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
